requirements.txt
streamlit>=1.31.0
Pillow>=10.0.0
ollama>=0.1.6
python-dotenv>=1.0.0 